refactored buffer conversion
