+++ apps/nextjs/src/components/createpost.tsx
@@ -7,6 +7,7 @@ import { useForm } from "react-hook-form";
 import { api } from "~/utils/api";
 import cross from "../../assets/cross.svg";
 import styles from "../styles/createpost.module.css";
+import { inputToBuffer } from "../utils/imageUtils";
 import { Button } from "./ui/button";
 import {
   Card,
@@ -145,50 +146,11 @@ const CreatePost: React.FC<CreatePostProps> = ({ onClose }) => {
       input.type = "file";
       input.accept = "image/*";
 
-      // Wait for the file to be selected
-      const fileSelected = new Promise<File | null>((resolve) => {
-        input.onchange = (event) => {
-          const fileInput = event.target as HTMLInputElement;
-          resolve(fileInput?.files?.[0] ?? null);
-        };
-        input.click();
-      });
-
-      const file = await fileSelected;
-
-      if (file) {
-        const reader = new FileReader();
-
-        // Return a promise that resolves when the file is read
-        const fileRead = new Promise<string>((resolve, reject) => {
-          reader.onloadend = () => resolve(reader.result as string);
-          reader.onerror = (error) => reject(error);
-          reader.readAsDataURL(file); // Read file as base64
-        });
-
-        const base64 = await fileRead;
-
-        if (base64) {
-          // Check if base64 contains a valid data URL format
-          const base64Data = base64.split(",")[1]; // Extract the base64-encoded string
-
-          if (base64Data) {
-            // Convert Base64 to binary
-            const binary = atob(base64Data);
-
-            // Convert binary to Buffer for S3 upload
-            const buffer = new Uint8Array(
-              binary.split("").map((char) => char.charCodeAt(0)),
-            );
-
-            // Example: Storing in a temporary array
-            setImagesTemp((prev) => [...prev, buffer]);
-          } else {
-            console.error("Base64 string is not valid.");
-          }
-        } else {
-          console.error("Error reading file");
-        }
+      const buffer = await inputToBuffer(input);
+      if (buffer) {
+        setImagesTemp((prev) => [...prev, buffer]);
+      } else {
+        console.error("Error with buffer.");
       }
     } catch (error) {
       console.error("Error picking image:", error);
